DEV: Implement proofParens and proofCollisions
proofParens encapsulates the check for files starting with a paren,
but not matching the correct template. It's used now both in the
folder Open and Reload button actions

proofCollisions encapsulates the check for colliding filenames.
It also is used in both the folder Open and Reload button actions.

The controls-enabled setters are now updated to disallow file
manipulations if any name collisions are found. This fixes #7.

This and previous commits likely fix #23.

The beginnings of an every-button hash check are also created here.