DEV: Roll out hash check; fix lbx population
Folder contents hash check rolled out to all include/exclude buttons,
as well as the 'generate spreadsheet' button. Warning is posted and form
buttons are disabled until user refreshes the folder.
The hash is updated after all actions that affect the files in the
folder (which is most of them: moves, insert, append, remove).

Closes #2, closes #15.

Listboxes had been set up to populate with anything matching the looser
regex constraint. Is now tighened so that they only populate if
rxInclFnameDetail or rxExclFnameDetail match them.