DEV: Implement variable pad width
Padding width is set with spinbutton and label near the top of the
form.  Default value is stored in code; the last-selected value
is stored in a custom DocumentProperty on the addin xlam.

In order for the form population to proceed correctly, the padding
*has* to be wide enough that all of the items naturally sort in
ascending order. Otherwise, packNums re-orders incorrectly, then
padNums shifts things around, and so on in an infinite loop. So,
popLists checks for a too-narrow padding before calling padNums
and adjusts it accordingly before proceeding.

packNums is updated to use the dynamically-selected padding width.

Various helper methods added (minPadWidth, setPadWidth,
getCustDocProp, setCustDocProp).

BUGFIX: If an 'All' button is clicked without anything being selected
in the 'source' listbox for that button. Crazy-bad things would
happen, from infinite loops to weird misbehavior.

Closes #20.