DEV: Update 'docstrings'; bit of cleanup
Docstrings are now freshened, including the one for
hashFilenames, which was simply wrong. Closes #25.

Other code comments also added/revised.

Remove unused 'collStr' and related code from checkNameCollisions.

Relocate addin file save to form unload, so as to speed the
workings of the form. Introduces a slight chance of loss of state
if an error occurs and the form drops without exit being called,
but this should hopefully happen rarely.

Helper subs added to clear the custom doc props (only one, for now)
and to rename the addin via BuiltinDocProps(1),
to facilitate creation of release binaries.